refactor : delete 시 조건문 추가

+++ src/main/java/roomescape/time/TimeService.java
@@ -3,6 +3,7 @@ package roomescape.time;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.error.ReferDataDeleteException;
 
 @Transactional
 @Service
@@ -13,14 +14,14 @@ public class TimeService {
         this.timeRepository = timeRepository;
     }
 
-    public Long save(final TimeRequest timeRequest) {
+    public Time save(final TimeRequest timeRequest) {
         return timeRepository.save(timeRequest);
     }
 
     @Transactional(readOnly = true)
     public Time findById(final Long id) {
         return timeRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("일치하는 예약이 없습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("일치하는 시간이 없습니다."));
     }
 
     @Transactional(readOnly = true)
@@ -30,6 +31,9 @@ public class TimeService {
 
     public void deleteById(final Long id) {
         findById(id);
+        if (timeRepository.findBySameReferId(id).isPresent()) {
+            throw new ReferDataDeleteException("해당 시간으로 예약한 정보가 존재합니다.");
+        }
         timeRepository.deleteById(id);
     }
 }
